current status of site

+++ obspost/gdrive.py
@@ -11,16 +11,22 @@ import urllib.request as urllib
 import random
 import string
 import os
+import requests
+import shutil
+import logging
 
+logger = logging.getLogger(__name__)
 
+flags = None
 # If modifying these scopes, delete your previously saved credentials
 # at ~/.credentials/sheets.googleapis.com-python-quickstart.json
 SCOPES = 'https://www.googleapis.com/auth/drive'
-CLIENT_SECRET_FILE = 'client_secret.json'
+CLIENT_SECRET_FILE = '/home/beme/client_secret.json'
+#CLIENT_SECRET_FILE = '/home/beme/drive-python-quickstart.json'
 APPLICATION_NAME = 'Google Sheets API Python Quickstart'
 
 
-def get_credentials():
+def get_credentials(flags=None):
     """Gets valid user credentials from storage.
 
     If nothing has been stored, or if the stored credentials are invalid,
@@ -41,11 +47,10 @@ def get_credentials():
     if not credentials or credentials.invalid:
         flow = client.flow_from_clientsecrets(CLIENT_SECRET_FILE, SCOPES)
         flow.user_agent = APPLICATION_NAME
-        if flags:
-            credentials = tools.run_flow(flow, store, flags)
-        else: # Needed only for compatibility with Python 2.6
-            credentials = tools.run(flow, store)
-        print('Storing credentials to ' + credential_path)
+        #if flags:
+        credentials = tools.run_flow(flow, store, flags)
+        #else: # Needed only for compatibility with Python 2.6
+        #    credentials = tools.run(flow, store)
     return credentials
 
 def upload_file(folder_id, url, file_name):
@@ -55,40 +60,50 @@ def upload_file(folder_id, url, file_name):
     students in a sample spreadsheet:
     https://docs.google.com/spreadsheets/d/1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms/edit
     """
-
     #Create a temporary file
     randstr = ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(0,10))
     fname = "/tmp/" + randstr + ".jpg"
-
     #Download the file at url into the temporary file
-    img_file = urllib.URLopener()
-    img_file.retrieve(url, fname)
+    r = requests.get(url, stream = True)
+    logger.info("URL is %s"%url)
+    logger.info("status code is %s"%str(r.status_code))
+    if r.status_code == 200:
+        logger.info("downloaded the picture from %s"%url)
+        with open(fname, 'wb') as f:
+            r.raw_decode_content = True
+            shutil.copyfileobj(r.raw, f)
+      
+#    img_file = urllib.URLopener()
+#    img_file.retrieve(url, fname)
 
-    #Create the MediaFileUpload Object
-    media = MediaFileUpload(fname, mimetype='image/jpg')
+        #Create the MediaFileUpload Object
+        media = MediaFileUpload(fname, mimetype='image/jpg')
 
-    #Obtain credentials for upload of file
-    credentials = get_credentials()
-    http = credentials.authorize(httplib2.Http())
-    service = discovery.build('drive', 'v3', http=http)
+        #Obtain credentials for upload of file
+        credentials = get_credentials()
+        http = credentials.authorize(httplib2.Http())
+        service = discovery.build('drive', 'v3', http=http)
 
-    file_metadata = {
-        'file_name': file_name,
-        'parents': [folder_id]
-    }
+        file_metadata = {
+            'file_name': file_name,
+            'parents': [folder_id]
+        }
 
-    # Execute the upload
-    afile = service.files().create(body=file_metadata, media_body=media,
-                                   fields='id').execute()
+        # Execute the upload
+        afile = service.files().create(body=file_metadata, media_body=media,
+                                       fields='id').execute()
 
-    file_id = afile.get('id')
-    file_update_metadata = {
-        'title': file_name
-    }
-    v2service = discovery.build('drive', 'v2', http=http)
-    updated_file = v2service.files().patch(fileId=file_id, body=file_update_metadata, fields='title').execute()
+        file_id = afile.get('id')
+        file_update_metadata = {
+            'title': file_name
+        }
+        v2service = discovery.build('drive', 'v2', http=http)
+        updated_file = v2service.files().patch(fileId=file_id, body=file_update_metadata, fields='title').execute()
 
-    #Return the id of the created file
+        #Return the id of the created file
 
-    os.remove(fname)
-    return file_id
+        os.remove(fname)
+        return file_id
+    else:
+        logger.error("The file was not created")
+        return None
